🏗️ Architect: Modularized SolarObjects
Modularized the `apts/objects/solar_objects.py` God Object by extracting its logic into a specialized `apts/objects/solar/` package.

- Created `apts/objects/solar/mpc.py` for MPC packed format parsing and minor planet ephem body configuration.
- Created `apts/objects/solar/calculations.py` for bulk Skyfield and Ephem computation loops.
- Created `apts/objects/solar/core.py` for the refactored `SolarObjects` class.
- Extracted `_to_float` utility to `apts/utils/math.py` for general reuse.
- Updated all internal imports and maintained backward compatibility via `apts/objects/__init__.py`.
