+++ apts/objects/__init__.py
@@ -1,6 +1,6 @@
 from .base import Objects as Objects
 from .messier import Messier as Messier
-from .solar_objects import SolarObjects as SolarObjects
+from .solar import SolarObjects as SolarObjects
 from .ngc import NGC as NGC
 from .stars import Stars as Stars
