+++ apts/objects/solar/__init__.py
@@ -1,0 +1,3 @@
+from .core import SolarObjects as SolarObjects
+
+__all__ = ["SolarObjects"]
