+++ tests/unit/test_solar_objects.py
@@ -1,7 +1,7 @@
 import unittest
 from datetime import datetime, timezone
 from tests import setup_southern_observation
-from apts.objects.solar_objects import SolarObjects
+from apts.objects import SolarObjects
 from apts.place import Place
 from apts.constants import ObjectTableLabels
 from apts.conditions import Conditions
